No more creating dummy task when buffer is empty, and resolved non correct int operations

+++ inf8480_tp2/src/main/java/com/inf8480_tp2/server/ComputeServerImpl.java
@@ -97,7 +97,7 @@ public class ComputeServerImpl implements ComputeServer {
         } else if (taskSize > 5 * this.serverCapacity) {
             return 100;
         } else {
-            return ((taskSize - this.serverCapacity) / (4 * this.serverCapacity)) * 100;
+            return ((float)(taskSize - this.serverCapacity) / (float)(4 * this.serverCapacity)) * 100;
         }
     }
